typedef changes. adding ca index for slist length store

+++ tests/test_bits.c
@@ -21,7 +21,7 @@ void test_cjks_htoni() {
 void test_read_be2() {
     char buf[] = { 0x0, 0x1 };
     cjks_io* io = cjks_io_mem_new(buf, 2);
-    unsigned short s = cjks_io_read_be2(io);
+    uint16 s = cjks_io_read_be2(io);
     assert(s == 1);
     cjks_io_mem_free(io);
 }
@@ -29,7 +29,7 @@ void test_read_be2() {
 void test_read_be4() {
     char buf[] = { 0x0, 0x0, 0x0, 0x1 };
     cjks_io* io = cjks_io_mem_new(buf, 4);
-    unsigned int i = cjks_io_read_be4(io);
+    uint32 i = cjks_io_read_be4(io);
     assert(i == 1);
     cjks_io_mem_free(io);
 }
@@ -37,7 +37,7 @@ void test_read_be4() {
 void test_read_be8() {
     char buf[] = { 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x1 };
     cjks_io* io = cjks_io_mem_new(buf, 8);
-    unsigned long long l = cjks_io_read_be8(io);
+    uint64 l = cjks_io_read_be8(io);
     assert(l == 1);
     cjks_io_mem_free(io);
 }
